Correct placement of changes
Fixes the order in view topics. Before any new topic created is shoved to the bottom _EDIT _ Actually is put in a random place depending on the topic starters user id.
as the last_post_id is set to the CURUSER['id'] on creating a new topic . Not sure if a fix could be changed there really. But this fix joins the tables and then sorts the view topics page by last post ADDED date. Seems to work but not fully tested mate.