remove pin call
Having the '$pin' value set to 'true' results in a never-ending loop of auth calls. Idle on an auth call results in rejection.

the deeper issue appears to be a broken $pin function in class_check.php, but setting this to ignore the $pin results in equal security settings as found in $memcache, etc.